Dashboard/requirements.txt
anyio==2.0.2
appnope==0.1.2
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
Babel==2.9.0
backcall==0.2.0
bleach==1.5.0
blis==0.7.4
cachetools==4.2.1
catalogue==2.0.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
click==7.1.2
contextvars==2.4
cymem==2.0.5
dataclasses==0.8
decorator==4.4.2
defusedxml==0.6.0
Django==3.1.4
django-heroku==0.0.0
entrypoints==0.3
enum34==1.1.6
filelock==3.0.12
ftfy==5.9
gunicorn==20.0.4
html5lib==0.9999999
httplib2==0.19.0
idna==2.10
immutables==0.14
importlib-metadata==3.4.0
jedi==0.18.0
Jinja2==2.11.2
joblib==1.0.1
json5==0.9.5
jsonschema==3.2.0
Markdown==2.6.9
MarkupSafe==1.1.1
mistune==0.8.4
murmurhash==1.0.5
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.4.3
nltk==3.4.5
notebook==6.2.0
numpy==1.19.5
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pathy==0.4.0
pexpect==4.8.0
pickleshare==0.7.5
preshed==3.0.5
prometheus-client==0.9.0
prompt-toolkit==3.0.10
protobuf==3.15.6
psycopg2==2.7.7
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.7.3
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2020.4
pyzmq==21.0.1
qtconsole==5.0.1
QtPy==1.9.0
regex==2020.11.13
requests==2.25.1
rsa==4.7.2
sacremoses==0.0.43
Send2Trash==1.5.0
six==1.15.0
smart-open==3.0.0
sniffio==1.2.0
sqlparse==0.4.1
srsly==2.4.0
terminado==0.9.2
testpath==0.4.4
thinc==8.0.1
tokenizers==0.9.4
tornado==6.1
tqdm==4.58.0
traitlets==4.3.3
transformers==4.2.2
typer==0.3.2
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
wasabi==0.8.2
wcwidth==0.2.5
Werkzeug==0.12.2
widgetsnbextension==3.5.1
whitenoise==5.2.0
zipp==3.4.0
